MFC r302507: Avoid truncating the return value of DTrace predicates.

+++ sys/cddl/contrib/opensolaris/uts/common/dtrace/dtrace.c
@@ -6336,7 +6336,7 @@ dtrace_probe(dtrace_id_t id, uintptr_t arg0, uintptr_t arg1,
 
 		if (pred != NULL) {
 			dtrace_difo_t *dp = pred->dtp_difo;
-			int rval;
+			uint64_t rval;
 
 			rval = dtrace_dif_emulate(dp, &mstate, vstate, state);
